+++ bin/main.dart
@@ -1,5 +1,0 @@
-import 'package:flutter_generate/flutter_generate.dart' as flutter_generate;
-
-main(List<String> arguments) {
-  print('Hello world: ${flutter_generate.calculate()}!');
-}