+++ example/lib/blocs/authentication/authentication_event.dart
@@ -1,0 +1,6 @@
+import 'package:equatable/equatable.dart';
+
+abstract class AuthenticationEvent extends Equatable {
+  const AuthenticationEvent();
+}
+