+++ example/lib/blocs/authentication/bloc.dart
@@ -1,0 +1,4 @@
+export 'authentication_bloc.dart';
+export 'authentication_event.dart';
+export 'authentication_state.dart';
+