+++ example/lib/repositories/authentication.dart
@@ -1,0 +1,7 @@
+
+class Authentication {
+
+  Authentication();
+
+}
+