+++ example/lib/widgets/sign_up_card.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class SignUpCard extends StatefulWidget {
+
+  @override
+  _SignUpCardState createState() => _SignUpCardState();
+}
+
+class _SignUpCardState extends State<SignUpCard> {
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
+