+++ lib/flutter_generate.dart
@@ -1,3 +1,3 @@
-int calculate() {
-  return 6 * 7;
-}
+export 'package:flutter_generate/generator.dart';
+export 'package:flutter_generate/structure.dart';
+export 'package:flutter_generate/file_model.dart';