+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 0.1.0
 
-- Initial version, created by Stagehand
+- Able to generate [bloc](https://pub.dev/packages/bloc) from commandline(with using [equatable](https://pub.dev/packages/equatable) only)
+
+- Initial release 
+- dartdoc Comments added
+
+- fix dependencies 
+- example and license created
+- Initial version, created